requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
base58==2.1.1
billiard==3.6.4.0
bitarray==2.6.2
celery==5.2.7
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cytoolz==0.12.1
decorators==2.0.7
Django==4.1.5
django-celery-beat==2.4.0
django-cors-headers==3.13.0
django-cron==0.6.0
django-rest-framework==0.1.0
django-timezone-field==5.0
djangorestframework==3.14.0
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.1
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
frozenlist==1.3.3
hexbytes==0.3.0
idna==3.4
ipfshttpclient==0.8.0a2
jsonschema==4.17.3
kombu==5.2.4
lru-dict==1.1.8
multiaddr==0.0.9
multidict==6.0.4
netaddr==0.8.0
openai==0.26.0
parsimonious==0.8.1
prompt-toolkit==3.0.36
protobuf==3.19.5
pycryptodome==3.16.0
pyrsistent==0.19.3
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.1
rlp==2.0.1
six==1.16.0
sqlparse==0.4.3
toolz==0.12.0
tqdm==4.64.1
tzdata==2022.7
urllib3==1.26.13
varint==1.0.2
vine==5.0.0
wcwidth==0.2.5
web3==5.31.3
websockets==9.1
yarl==1.8.2
